requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
transformers==4.35.2
torch>=2.0.0
pillow>=10.0.0
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
requests>=2.31.0
# Enhanced dependencies (simplified)
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
# spacy>=3.7.0  # Commented out due to compilation issues
# accelerate>=0.24.0  # Commented out due to compilation issues
